requirements.txt
# Python packages
lxml
playwright
bs4b